+++ common/Cargo.toml
@@ -11,6 +11,7 @@ harness = false
 anyhow.workspace = true
 crc32fast.workspace = true
 image.workspace = true
+itertools.workspace = true
 nalgebra.workspace = true
 serde.workspace = true
