+++ slicer/src/post_process/mod.rs
@@ -1,3 +1,2 @@
-pub mod downsample;
 pub mod elephant_foot_fixer;
 pub mod island_detection;
